SPORT-14 Add implementation for having ability to repeat workout by getting the latest previous data

+++ App/Workout/Services/WorkoutService/WorkoutQueryService.cs
@@ -39,7 +39,7 @@ namespace SaveApp.App.Workout.Services.WorkoutService
                 WorkoutExercise workoutExerciseFromDb =
                     _queryRepository.GetLatestWorkoutExerciseById(
                         userId,
-                        workoutExercise.Id,
+                        (int)workoutExercise.Id,
                         workoutExercise.Exercise.Id
                     );
